+++ src/features/board/BoardView.tsx
@@ -8,9 +8,9 @@ import { NavButtons } from "./navigation/NavButtons";
 import { useBoardNavigation } from "./navigation/useBoardNavigation";
 import { SuggestionBar } from "./suggestions/SuggestionBar";
 import { useSuggestions } from "./suggestions/useSuggestions";
-import { Tile } from "./Tile";
+import { Tile } from "./tile/Tile";
+import { useTileActivation } from "./tile/useTileActivation";
 import type { Board, BoardButton } from "./types";
-import { useTileActivation } from "./useTileActivation";
 
 export interface BoardViewProps {
   board: Board;
