+++ src/features/board/pictogram/Pictogram.test.tsx
