+++ src/features/board/pictogram/Pictogram.tsx
